+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Minor
 
 - Tooltip: Show tooltip when React children have focus (#824)
+- IconButton/Pog: Removed unused iconColor options: blue, orange; added darkGray; added bgColor: darkGray (#823)
+- Docs: Replaced combinations in Pog with Combinations: Icon Color & Background Color. Removed IconButton-bgColor-blue option from Docs. (#823)
 
 ### Patch